:tada: introduce :filter and :update_expressions
This commit introduces :filter (instead of :query-key and :query-value) for
defining the query that we pass down to Mongo. It is a hash and will allow us
to filter by a composite key.

In addition to that :update_expressions is now another optional hash that can
be added if the action is an update.

The first key/value can be a Mongo Update
Expression (https://docs.mongodb.com/manual/reference/operator/update/#id1) and
the values substituted as usual. Note that pipeline (Mongo >= 4.2) support is
not there yet.

The :update_expression hash will *replace* the $set operator that includes the
logstash event itself.

On top of it all, now action is fully dynamic and expanded via sprintf as in
logstash-output-elasticsearch.